Move existing Micronaut 2.+ instrumentation into a dedicated folder
Extract MicronautDecorator into a common folder
