+++ dd-java-agent/instrumentation/micronaut/http-server-netty-2.0/src/test/resources/application.properties
