+++ settings.gradle
@@ -254,7 +254,8 @@ include ':dd-java-agent:instrumentation:log4j-2.7'
 include ':dd-java-agent:instrumentation:log4j1'
 include ':dd-java-agent:instrumentation:log4j2'
 include ':dd-java-agent:instrumentation:logback-1'
-include ':dd-java-agent:instrumentation:micronaut-http-server-netty-2'
+include ':dd-java-agent:instrumentation:micronaut'
+include ':dd-java-agent:instrumentation:micronaut:http-server-netty-2.0'
 include ':dd-java-agent:instrumentation:mongo'
 include ':dd-java-agent:instrumentation:mongo:common'
 include ':dd-java-agent:instrumentation:mongo:bson-document'
